move: contentful 관련 유틸 관리 폴더 이동(lib -> utils
